refactor cloud account overview classes to simplify logic and better use built-in DRF serializer functionality
This also changes the behavior of the API when authenticatd as a superuser.
Previously, as a superuser it would return **all** accounts in the system,
but that's a huge performance risk and is inconsistent with how we defined
the other API that powers the dashboard page. Now, the superuser sees her
own accounts, just as if she was a regular user. Filtering to an optional
user_id when authenticated as superuser remains unchanged.