refactor to make images usable by multiple users
This commit includes some unfortunately major changes in order to support
the concept of multiple users and accounts having shared access to the
underlying images, but this brings us a much more accurate representation
to how AWS treats images.

We now no longer have specific accounts in our system assigned directly
to each image because we may have to handle images for accounts that we
do not have in our system. Previously, we wrongly assumed that the account
that discovered the image was the owner of the image. Instead, AWS images
now have a new property to indicate the owner AWS account ID.

This commit changes some of the initial inspection and CloudTrail log
analysis processes because we need to describe the AWS AMI in order to
know what AWS account ID is its owner.

When a user of the cloudigrade API interacts with images, that user now
has access to all images that have been referenced by any event for any
instance under any account belonging to that user. This means that
multiple users can now see the same underlying images, but they will not
see images which which they have never interacted.