Change recalculate_runs logic-
Instead of getting relevant events in task and passing to recalculate runs,
have recalculate_runs only get the one event. Based on the event occuring time
and the relevant runs, normalize and recalculate.

This is prefered over the previous logic, since the previous implementation
changed what runs are generated based on the order of events passed-
and we can easily end up with runs with no endtimes for a spurious start event.