explicitly define names for all celery tasks
Note that the tasks under the aws.tasks package are named as if their
functions live in the aws.tasks module itself instead of the next-
level modules (e.g. aws.tasks.cloudtrail). This is because we moved
those functions in e4abbbc, but because we were using automatic naming
and we had existing task instances in the database and queues, when
Celery tried running those old tasks with the new code, it exploded
spectacularly with lots of errors.

Henceforth, we should always define explicit names on our tasks to
protect ourselves from encountering this type of problem again with
future code refactoring efforts.