save calculation task while we have the user lock
this should handle the edge case where we release
the lock and then the user is swiftly deleted.
that could happen if there was a separate delete
transaction blocked until we released the lock.
if that happened with the old code, we would try
to save and update the calculation task after it
was deleted (as a result of the delete cascade
from the user), and that would result in a db
integrity error.