add psql cli tools to cloudigrade image
I'm adding these to help diagnose some issues in our deployed
environments. This might be temporary, but it might also be useful
to keep around for future investigations.

Yes, these curls and rpm commands are kind of ugly, but neither UBI
nor EPEL include appropriate postgresql packages. Also, we have to
tell rpm "--nodeps" because we don't want it to pull in systemd and
all of *its* dependencies (which postgresql11 declares but really
only needs if you want to run the psql *server*).