Bumping up kombu to 5.2.4, we need to bring in the fix from https://github.com/celery/kombu/pull/1495
This fixes the issue where task events from the cloudigrade
worker were not seen by the internal celery metrics.

High level:
With kombu 5.2.3, redis client psubscribes are not subscribed
to a channel named with the global_keyprefix that we
specified with the broker_transport options. However, task
and worker events are keyed with the prefix, so they are never
seen by the celery metrics. Removing the global_keyprefix
(i.e. user prefix via CLOUDIGRADE_ENVIRONMENT)
events now are properly seen, but this won't work with a shared
redis instance.

With the fix in the PR mentioned above, the global_keyprefix
is properly handled with pubsub clients.