make get_or_create_user handle legacy users that have no org_id
If we had an existing api.User with no org_id but called get_or_create_user with an org_id, the previous code would erroneously try to create a new api.User because it didn't expect that they could exist with no org_id. This change allows org_id to be missing during the get part of get_or_create.