restore "username" filtering on internal users api
We removed the ability to filter the internal "users" API by the "username" query string argument in https://github.com/cloudigrade/cloudigrade/pull/1192, but we recently learned that the swatch service was relying on that. This change reimplements that filter by filtering on User.account_number with the given "username" and adds several unit tests to ensure we don't accidentally lose this in the future.

* allow filtering internal users API by "username" (which actually filters "account_number")
* add "username" to internal users API responses
* allow filtering internal users API by "uuid"